Fula drive (#49)
Add FulaDrive on top of FxFS

Having the `FxFS` as a filesystem based on UnixFS, the next layer placed on top of the fs layer is FulaDrive.
FulaDrive wraps the `FxFS` layer and provides a drive-like interface that hides all of the complexity brought over by immutability in the `FxFS`.
Each drive has two different DriveSpaces (PublicSpace and PrivateSpace). Separating these two will enable us to store different file types on them (File and EncodedFile).
DriveStore interface is added for handling the immutability. Every change in the drive will change the DAG leading to a change in CIDs. Publishing a FulaDrive on a DriveStore will update the references to root directories.

The interface is intended to be compatible with POSIX comprising:
- `WriteFile` for writing a file on the drive
  -- on the PrivateSpace it will also write a JWE object alongside the file
- `DeleteFile` for deleting a file from the drive
- `ReadFile` for reading a file from the drive
  -- on the PrivateSpace it will also read a JWE object alongside the file
- `ListEntries` for listing entries inside a directory in the drive

Tests for each of the methods are added to ensure every added feature integrates with the previous correctly.