Newfile protocol (#64)
NewFile protocol is the next version of the File Protocol. It will get renamed after all tests are passed on the `mobile` package.

The new version offers a POSIX-compatible interface. Every file request gets initiated with an `FSRequest` message. An `FSRequest` asks for a particular `action` on a particular `user`'s drive. After sending the `FSRequest` the requesting side will send a `payload` if necessary. Right now, the only action that requires a payload to be sent is `Write` (the payload being the file content).
To have a multi-phase process (sending a request and after that sending payload), the requester must send the size of the `FSRequest` message (in a 2-bytes length message) before sending the `FSRequest` itself.

The actions are:
- `Write`: request to write a file in a user's drive
- `Read`: request to read a file from a user's drive
- `Delete`: request to delete a file from a user's drive
- `Ls`: request a list of entries in a directory inside a user's drive
- `MkDir`: request to create a new directory in a user's drive