File Protocol Replaced(#68)
The previous version of the file protocol is removed and replaced with the new one (explained in #64)

Dependencies are fixed, meaning that `kubo` and `libp2p` versions are set to a certain point in their repositories, all tests are passed with these versions. Also dependencies to `functionland/go-unixfs` and `functionland/go-mfs` are fixed.

The mobile module is updated to work with the new version of the File Protocol. Interfaces are changed to match the language of the new protocol. Some methods like `NewFula` are kept the same as old.