Remove redundant packages
A series of package are now made redundant in this repo in light of
using WNFS. They make up the majority of code in this repo, including:
`protocol/file`, `drive` and `fxfs`.

There are a set of dependant and unused packages that are also removed
as a result of this clean up, which include: `beast`, `crypto` and
`common`.

The `mobile` package needs to be completely re-implemented which is also
 removed as part of this commit.