Enable auto syncing of DAG on `Get` and `Put`
Configure fluamobile client such that on `Get` it automatically pulls
the missing links and on `Put` it automatically pushes the links to
Blox.

Note, the Push and Pull API exposed by the client is changed to take
addr as a config, because signature of `Get` and `Put` cannot change to
take it due to WNFS store API contract.

The changes here also fix the type conversions in gomobile binding of
Java.