Implement noop DAG exchange for testing purposes
Introduce `noop` exchange, where the operations are only logged and no
remote connection is made. This is to enable testing without having to
have an accessible Blox API running.

A new config is also introduced in `fulamobile` package called
`Exchange`, where if the value is set to `noop` the remote exchange is
skipped and `BloxAddr` config may be unspecified.