Added methods to store and fetch by identity
We want to be able to store root Cid of encrypted hamt tree from WNFS in the backend. But it should be retrievable only with identiy (which is something we can recreate).

We added two methods for store and get. One receives the encrypted versio of root Cid and returns the encrypted version and JAVA code must handle the encryption and decryption, and one receives a key and handles the decryption and encryption itself